Working CountVectorizer, TfidfTransformer Pipeline
A very rough but working script that gets ~60% accuracy for BernoulliNB, 69% for k-NN, 75-77% for Multinomial NB, SGDClassifier, Logistic Regression, and Perceptron.  We need to generalize the file so it can take other data sets, as well as loop over all models so we can compare side by side. We also want to look at Pipelines, preprocessing (already tried english stopwords and ngrams to no benefit) and GridSearchCV to find best parameters for a given model.